@@ -1,6 +1,6 @@
 {
   "name": "angular-starter",
-  "version": "10.0.14",
+  "version": "10.1.1",
   "scripts": {
     "ng": "ng",
     "start": "ng serve --port 4200",
@@ -15,7 +15,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/ganatan/angular-example-pwa.git"
+    "url": "https://github.com/ganatan/angular-example-seo.git"
   },
   "author": {
     "name": "danny",
@@ -24,43 +24,43 @@
   },
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/ganatan/angular-example-pwa/issues"
+    "url": "https://github.com/ganatan/angular-example-seo/issues"
   },
-  "homepage": "https://github.com/ganatan/angular-example-pwa#readme",
+  "homepage": "https://github.com/ganatan/angular-example-seo#readme",
   "private": true,
   "dependencies": {
-    "@angular/animations": "10.0.14",
-    "@angular/common": "10.0.14",
-    "@angular/compiler": "10.0.14",
-    "@angular/core": "10.0.14",
-    "@angular/forms": "10.0.14",
-    "@angular/platform-browser": "10.0.14",
-    "@angular/platform-browser-dynamic": "10.0.14",
-    "@angular/platform-server": "10.0.14",
-    "@angular/router": "10.0.14",
-    "@angular/service-worker": "10.0.14",
+    "@angular/animations": "10.1.1",
+    "@angular/common": "10.1.1",
+    "@angular/compiler": "10.1.1",
+    "@angular/core": "10.1.1",
+    "@angular/forms": "10.1.1",
+    "@angular/platform-browser": "10.1.1",
+    "@angular/platform-browser-dynamic": "10.1.1",
+    "@angular/platform-server": "10.1.1",
+    "@angular/router": "10.1.1",
+    "@angular/service-worker": "10.1.1",
     "@fortawesome/fontawesome-free": "5.14.0",
-    "@nguniversal/express-engine": "10.0.2",
+    "@nguniversal/express-engine": "10.1.0",
     "bootstrap": "4.5.2",
     "express": "4.17.1",
     "jquery": "3.5.1",
-    "rxjs": "6.6.2",
+    "rxjs": "6.6.3",
     "tslib": "2.0.1",
     "zone.js": "0.11.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.1000.8",
-    "@angular/cli": "10.0.8",
-    "@angular/compiler-cli": "10.0.14",
-    "@nguniversal/builders": "10.0.2",
-    "@types/express": "4.17.7",
-    "@types/node": "14.6.2",
+    "@angular-devkit/build-angular": "0.1001.1",
+    "@angular/cli": "10.1.1",
+    "@angular/compiler-cli": "10.1.1",
+    "@nguniversal/builders": "10.1.0",
+    "@types/express": "4.17.8",
+    "@types/node": "14.10.1",
     "@types/jasmine": "3.5.14",
     "@types/jasminewd2": "2.0.8",
     "codelyzer": "6.0.0",
     "jasmine-core": "3.6.0",
     "jasmine-spec-reporter": "5.0.2",
-    "karma": "5.1.1",
+    "karma": "5.2.2",
     "karma-chrome-launcher": "3.1.0",
     "karma-coverage-istanbul-reporter": "3.0.3",
     "karma-jasmine": "4.0.1",
@@ -68,6 +68,6 @@
     "protractor": "7.0.0",
     "ts-node": "9.0.0",
     "tslint": "6.1.3",
-    "typescript": "3.9.7"
+    "typescript": "4.0.2"
   }
 }
