@@ -23,11 +23,11 @@ export class ExerciceComponent implements OnInit {
     this.movie = new Movie();
   }
 
-  ngOnInit() {
+  ngOnInit(): void {
     this.updateControls();
   }
 
-  updateClass() {
+  updateClass(): void {
     this.movie.name = 'Avengers: Endgame';
     this.movie.releaseDate = '26/04/2019';
     this.movie.franchise = true;
@@ -37,7 +37,7 @@ export class ExerciceComponent implements OnInit {
       'the universe is in ruins.';
   }
 
-  updateControls() {
+  updateControls(): void {
     this.updateClass();
     this.name.setValue(this.movie.name);
     this.releaseDate.setValue(this.movie.releaseDate);
@@ -47,7 +47,7 @@ export class ExerciceComponent implements OnInit {
     this.summary.setValue(this.movie.summary);
   }
 
-  resetControls() {
+  resetControls(): void {
     this.name.setValue(null);
     this.releaseDate.setValue(null);
     this.franchise.setValue(null);
@@ -56,7 +56,7 @@ export class ExerciceComponent implements OnInit {
     this.summary.setValue(null);
   }
 
-  getClass() {
+  getClass(): void {
     this.movie.name = this.name.value;
     this.movie.releaseDate = this.releaseDate.value;
     this.movie.franchise = this.franchise.value;
