@@ -19,11 +19,11 @@ export class ExerciceComponent implements OnInit {
 
   constructor() { }
 
-  ngOnInit() {
+  ngOnInit(): void {
     this.updateControls();
   }
 
-  updateControls() {
+  updateControls(): void {
     this.exampleForm.patchValue({
       name: 'Avengers: Endgame',
       releaseDate: '26/04/2019',
@@ -35,7 +35,7 @@ export class ExerciceComponent implements OnInit {
     });
   }
 
-  resetControls() {
+  resetControls(): void {
     this.exampleForm.patchValue({
       name: null,
       releaseDate: null,
