@@ -24,11 +24,11 @@ export class ExerciceComponent implements OnInit {
 
   constructor(private fb: FormBuilder) { }
 
-  ngOnInit() {
+  ngOnInit(): void {
     this.updateControls();
   }
 
-  updateControls() {
+  updateControls(): void {
     this.exampleForm.patchValue({
       name: 'Avengers: Endgame',
       releaseDate: '26/04/2019',
@@ -44,7 +44,7 @@ export class ExerciceComponent implements OnInit {
     });
   }
 
-  resetControls() {
+  resetControls(): void {
     this.exampleForm.patchValue({
       name: null,
       releaseDate: null,
