@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { Observable } from 'rxjs';
 
 const httpOptions = {
   headers: new HttpHeaders(
@@ -16,7 +17,7 @@ export class ItemsService {
 
   constructor(private http: HttpClient) { }
 
-  getItems(url: string) {
+  getItems(url: string): Observable<object> {
     return this.http.get(url, httpOptions);
   }
 
