@@ -3,7 +3,7 @@ export const environment = {
   application:
   {
     name: 'angular-starter',
-    angular: 'Angular 10.0.14',
+    angular: 'Angular 10.1.1',
     bootstrap: 'Bootstrap 4.5.2',
     fontawesome: 'Font Awesome 5.14.0',
   }
